lint; refactor
